refactor: move strategy tests to a separate directory
